+++ src/bj.h
@@ -3,7 +3,7 @@
 /* declarations for the blackjack module in the casino */
 
 typedef vtype drawcard[16];
-typedef integer hand[16];
+typedef long hand[16];
 
 extern void bj__game_blackjack();
